+++ src/gui/gui-buffer.c
@@ -4511,10 +4511,12 @@ gui_buffer_dump_hexa (struct t_gui_buffer *buffer)
             free (tags);
         snprintf (buf, sizeof (buf), "%s", ctime (&ptr_line->data->date));
         buf[strlen (buf) - 1] = '\0';
-        log_printf ("  date:         %d = %s", ptr_line->data->date, buf);
+        log_printf ("  date:         %lld = %s",
+                    (long long)(ptr_line->data->date), buf);
         snprintf (buf, sizeof (buf), "%s", ctime (&ptr_line->data->date_printed));
         buf[strlen (buf) - 1] = '\0';
-        log_printf ("  date_printed: %d = %s", ptr_line->data->date_printed, buf);
+        log_printf ("  date_printed: %lld = %s",
+                    (long long)ptr_line->data->date_printed, buf);
 
         /* display raw message for line */
         if (ptr_line->data->message)
